loading/lazyload/image-loading-lazy-in-script-disabled-iframe-002.tentative.html

<!DOCTYPE html>
<head>
<title>Moving lazy-loading image into script disabled iframe</title>
<link rel="help" href="https://github.com/scott-little/lazyload">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<div style="height:10000px;"></div>
<iframe id="iframe" src="resources/image-loading-lazy-in-viewport-iframe.html">
</iframe>
<iframe id="sandboxediframe" sandbox="allow-same-origin"
        src="resources/subframe.html">
</iframe>
<script>
promise_test(async t => {
  const promise1 = new Promise(resolve => iframe.addEventListener("load", resolve));
  const promise2 = new Promise(resolve => sandboxediframe.addEventListener("load", resolve));
  await Promise.all([promise1, promise2]);

  const image = iframe.contentDocument.querySelector("img");

  assert_false(image.complete, "lazy-load image shouldn't be loaded");

  sandboxediframe.contentDocument.body.appendChild(image);

  await new Promise(resolve => image.addEventListener("load", resolve));

  assert_true(image.complete,
              "lazy-load image shouldn't be honored in script disabled iframe");
});
</script>
